+++ changelog.md
@@ -66,6 +66,7 @@
 * PayPal Standard: Costs for checkout attributes were double charged
 * Paging of return request grid did not work
 * #428 Multiline checkout attributes aren't rendered correctly
+* #434 Shipping.ByTotal: Make grid pageable
 
 
 ##SmartStore.NET 2.0.2#